+++ s3proxy/internal/router/BUILD.bazel
@@ -4,6 +4,7 @@ load("//bazel/go:go_test.bzl", "go_test")
 go_library(
     name = "router",
     srcs = [
+        "handler.go",
         "object.go",
         "router.go",
     ],
